requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.2
Django==5.1.7
dnspython==2.7.0
dnsrecon==0.10.1
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lxml==5.3.1
MarkupSafe==3.0.2
nassl==5.3.0
netaddr==1.3.0
nltk==3.9.1
numpy==2.2.2
nvdlib==0.8.0
pluginbase==1.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PySocks==1.7.1
python-nmap==0.7.1
pytz==2025.2
regex==2024.11.6
requests==2.32.3
setuptools==75.6.0
sqlparse==0.5.3
sslyze==6.1.0
theHarvester==0.0.1
tls_parser==2.0.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
wafw00f==2.3.1
Werkzeug==3.1.2
wheel==0.45.1
